+++ src/core.js
@@ -201,7 +201,7 @@ function injectCSSFallback() {
 }
 
 const Kineto = {
-  version: '0.8.26',
+  version: '0.8.28',
 
   get env() {
     if (!cachedEnv) cachedEnv = env();